+++ ui-app/app/layout.tsx
@@ -3,7 +3,7 @@ import localFont from "next/font/local";
 import "./globals.css";
 import { GlobalStateProvider } from "@/lib/GlobalStateContext";
 import { Toaster } from "react-hot-toast";
-import { ThemeProvider } from "@/components/theme-provider";
+import { ThemeProvider } from "@/components/core/theme-provider";
 
 const geistSans = localFont({
   src: "./fonts/GeistVF.woff",
