+++ ui-app/components/AccordionFormik.tsx
@@ -8,9 +8,9 @@ import { Formik, Form, ErrorMessage } from "formik";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Plus } from "lucide-react";
-import { Mandatory } from "@/components/Mandatory";
+import { Mandatory } from "@/components/core/Mandatory";
 import * as Yup from "yup";
-import { Tooltips } from "./Tooltips";
+import { Tooltips } from "./core/Tooltips";
 import { Textarea } from "./ui/textarea";
 import { FormikSelect } from "./FormikSelect";
 import Operators from "@/lib/operators";
